// src/types/Media.d.ts
export interface Media {
  id: number;
  model_type: string;
  model_id: number;
  uuid: string;
  collection_name: string;
  name: string;
  file_name: string;
  mime_type: string;
  disk: string;
  conversions_disk: string;
  size: 28213;
  // "manipulations": [],
  // "custom_properties": [],
  // "generated_conversions": [],
  // "responsive_images": [],
  order_column: 1;
  created_at: string;
  updated_at: string;
  original_url: string;
  preview_url: string;
}